혁재 : 이동동기화 수정, vec2 추가

+++ DevProject/ProjectQ_Server/ConsoleServer/GameServer/PlayerObject/PlayerData.cs
@@ -4,8 +4,8 @@ namespace GameServer.Player
 {
     public class PlayerData : IPlayerInfo
     {
-        public float Xpos { get; set; }
-        public float Ypos { get; set; }
+        public double Xpos { get; set; }
+        public double Ypos { get; set; }
         public int Level { get; set; }
         public int Exp { get; set; }
     }
